backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0
beautifulsoup4>=4.12.0
pydantic-settings>=2.5.0

langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=0.1.0
langchain-ollama>=0.2.0
huggingface_hub>=0.23.0
langchain-text-splitters>=0.3.0

faiss-cpu>=1.8.0
tiktoken>=0.7.0
openai>=1.50.0

pypdf>=5.0.0
python-docx>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10

sentence-transformers>=3.0.0

numpy>=1.26.0

ollama>=0.4.0
